+++ requirements.txt
@@ -2,7 +2,6 @@
 click>=8.0.0
 numpy>=1.21.0
 soundfile>=0.12.0
-pathlib
 requests>=2.28.0
 python-dotenv>=1.0.0
 
@@ -21,3 +20,4 @@ moviepy>=1.0.0
 pytest>=7.0.0
 black>=22.0.0
 flake8>=5.0.0
+types-requests==2.32.4.20250611
