+++ src/audio/dialogue/__init__.py
@@ -3,11 +3,11 @@ Audio Dialogue Module
 Enhanced dialogue generation and management for multi-speaker conversations
 """
 
-from .dialogue_base import DialogueBase, DialogueSegment
-from .dialogue_converter import TranscriptToDialogueConverter
-from .dialogue_generator import DialogueGenerator
-from .character_profile import CharacterProfile, CharacterManager
-from .dialogue_test import DialogueTester
+from .base import DialogueBase, DialogueSegment
+from .converter import TranscriptToDialogueConverter
+from .generator import DialogueGenerator
+from .profile import CharacterProfile, CharacterManager
+from .test import DialogueTester
 
 __all__ = [
     'DialogueBase',
