+++ src/audio/providers/assemblyai/__init__.py
@@ -2,6 +2,6 @@
 AssemblyAI provider for speaker diarization
 """
 
-from .ai_diarizer import AssemblyAIDiarizer
+from .diarizer import AssemblyAIDiarizer
 
 __all__ = ['AssemblyAIDiarizer']
