+++ src/audio/providers/elevenlabs/__init__.py
@@ -4,7 +4,7 @@ TTS Provider: Generates separate audio tracks from transcript segments using Ele
 STT Provider: Transcribes audio/video files with speaker diarization using ElevenLabs Scribe API
 """
 
-from .el_tts import ElevenLabsTTSProvider
-from .el_stt import ElevenLabsSTTProvider
+from .tts import ElevenLabsTTSProvider
+from .stt import ElevenLabsSTTProvider
 
 __all__ = ['ElevenLabsTTSProvider', 'ElevenLabsSTTProvider']
