+++ src/audio/providers/elevenlabs/stt.py
@@ -62,7 +62,7 @@ class ElevenLabsSTTProvider(SpeakerDiarizationProvider):
     def _get_api_key(self) -> Optional[str]:
         """Get API key from multiple sources"""
         # Load from .env.local file first
-        from ....utils.util_env import get_env
+        from ....utils.env import get_env
         
         # Try environment variables
         api_key = get_env('ELEVENLABS_API_KEY')
