+++ src/audio/providers/elevenlabs/tts.py
@@ -19,7 +19,7 @@ from elevenlabs import ElevenLabs
 from elevenlabs.types import VoiceSettings
 
 from ..base import SpeakerDiarizationProvider
-from .el_voice import VoiceManager
+from .voice import VoiceManager
 
 logger = logging.getLogger(__name__)
 
@@ -70,7 +70,7 @@ class ElevenLabsTTSProvider(SpeakerDiarizationProvider):
     def _get_api_key(self) -> Optional[str]:
         """Get API key from multiple sources"""
         # Load from .env.local file first
-        from ....utils.util_env import get_env
+        from ....utils.env import get_env
         
         # Try environment variables
         api_key = get_env('ELEVENLABS_API_KEY')
