+++ src/text/editor.py
