+++ src/workflows/__init__.py
@@ -3,6 +3,6 @@ Workflow modules for DuoSynco
 Orchestrates complex multi-step processing workflows
 """
 
-from .edit_workflow import EditWorkflow
+from .edit import EditWorkflow
 
 __all__ = ['EditWorkflow']
